+++ src/node/ext/ldap/main.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from __future__ import print_function
 from importlib import import_module
 from node.ext.ldap import testing
 import argparse
@@ -30,7 +31,7 @@ parser.add_argument(
     'ldiflayer',
     nargs='?',
     default='base',
-    choices=testing.ldif_layer.keys(),
+    choices=list(testing.ldif_layer.keys()),
     help='Predefined LDIF Layer to load.'
 )
 
@@ -115,7 +116,7 @@ def slapd():
         # XXX should check for distinct slapd
         checkslapd = 'ps ax| grep slapd| grep -v grep -q'
         if not subprocess.call(checkslapd, shell=True) == 1:
-            print u"LDAP already running. abort."
+            print(u"LDAP already running. abort.")
             sys.exit(2)
         startslapd(layer, ns.ldiflayer)
     else: