+++ src/node/ext/ldap/testing/ldifs/genSambaUsers.py
@@ -1,7 +1,9 @@
 #!/usr/bin/env python
 
+from __future__ import print_function
 import sys
 import smbpasswd
+from six.moves import range
 
 container_template = """\
 dn: ou=sambaUsers,dc=my-domain,dc=com
@@ -29,12 +31,12 @@ userPassword: %(pwd)s
 
 n = int(sys.argv.pop(1))
 
-print container_template
+print(container_template)
 for x in range(n):
     secret = 'secret%d' % x
     nt_secret = smbpasswd.nthash(secret)
     lm_secret = smbpasswd.lmhash(secret)
-    print user_template % dict(
+    print(user_template % dict(
             uid='uid%d' % x,
             cn='cn%d' % x,
             uid_num=str(x),
@@ -44,4 +46,4 @@ for x in range(n):
             pwd=secret,
             ntpwd=nt_secret,
             lmpwd=lm_secret,
-            ),
+            ), end=' ')