+++ src/node/ext/ldap/tests.py
@@ -1,9 +1,12 @@
 # -*- coding: utf-8 -*-
+from __future__ import print_function
 from node.ext.ldap import testing
 from plone.testing import layered
 import doctest
 import interlude
 import pprint
+import re
+import six
 import unittest
 
 
@@ -29,12 +32,32 @@ optionflags = \
     doctest.REPORT_ONLY_FIRST_FAILURE
 
 
-print """
+print("""
 *******************************************************************************
 If testing while development fails, please check if memcached is installed and
 stop it if running.
 *******************************************************************************
-"""
+""")
+
+
+class Py23DocChecker(doctest.OutputChecker):
+    exclude = [
+        #"<.*>$",  # object repr
+        #"^\S*Error: ",  # raised error
+    ]
+
+    def check_output(self, want, got, optionflags):
+        if want != got and six.PY2:
+            # if running on py2, ignore any "u" prefixes in the output
+            if not any(re.match(pattern, want) for pattern in self.exclude):
+                got = re.sub("(\\W|^)u'(.*?)'", "\\1'\\2'", got)
+                got = re.sub("(\\W|^)u\"(.*?)\"", "\\1\"\\2\"", got)
+            # also ignore "b" prefixes in the expected output
+            want = re.sub("b'(.*?)'", "'\\1'", want)
+            # we get 'ldap.' prefixes on python 3, e.g.
+            # ldap.UNWILLING_TO_PERFORM
+            want = want.lstrip('ldap.')
+        return doctest.OutputChecker.check_output(self, want, got, optionflags)
 
 
 def test_suite():
@@ -48,6 +71,7 @@ def test_suite():
                        'pp': pprint.pprint,
                        },
                 optionflags=optionflags,
+                checker=Py23DocChecker(),
             ),
             layer=layer)
         for docfile, layer in DOCFILES