+++ src/node/ext/ldap/ugm/defaults.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
-import posix
-import samba
+from node.ext.ldap.ugm import posix
+from node.ext.ldap.ugm import samba
 
 # required default callbacks for several object classes.
 creation_defaults = dict()